refactor to keep all changes in core
