+++ janusgraph-core/src/main/java/org/janusgraph/diskstorage/dlq/DLQManager.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package org.janusgraph.diskstorage.es.dlq;
+package org.janusgraph.diskstorage.dlq;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
