+++ janusgraph-core/src/main/java/org/janusgraph/diskstorage/indexing/IndexTransaction.java
@@ -19,7 +19,11 @@ import org.janusgraph.diskstorage.BackendException;
 import org.janusgraph.diskstorage.BaseTransaction;
 import org.janusgraph.diskstorage.BaseTransactionConfig;
 import org.janusgraph.diskstorage.LoggableTransaction;
+import org.janusgraph.diskstorage.configuration.Configuration;
+import org.janusgraph.diskstorage.dlq.DLQManager;
+import org.janusgraph.diskstorage.dlq.ElasticSearchDLQ;
 import org.janusgraph.diskstorage.util.BackendOperation;
+import org.janusgraph.graphdb.configuration.GraphDatabaseConfiguration;
 import org.janusgraph.graphdb.database.idhandling.VariableLong;
 import org.janusgraph.graphdb.database.serialize.DataOutput;
 import org.janusgraph.graphdb.tinkerpop.optimize.step.Aggregation;
@@ -32,6 +36,10 @@ import java.util.List;
 import java.util.Map;
 import java.util.concurrent.Callable;
 import java.util.stream.Stream;
+
+import static org.janusgraph.graphdb.configuration.GraphDatabaseConfiguration.DLQ_ENABLED;
+import static org.janusgraph.graphdb.configuration.GraphDatabaseConfiguration.DLQ_KAFKA_BOOTSTRAP_SERVERS;
+import static org.janusgraph.graphdb.configuration.GraphDatabaseConfiguration.DLQ_KAFKA_TOPIC;
 
 /**
  * Wraps the transaction handle of an index and buffers all mutations against an index for efficiency.
@@ -54,8 +62,13 @@ public class IndexTransaction implements BaseTransaction, LoggableTransaction {
 
     private Map<String,Map<String,IndexMutation>> mutations;
 
+    // Dead Letter Queue
+    private final ElasticSearchDLQ dlq;
+    private final String indexName;
+
     public IndexTransaction(final IndexProvider index, final KeyInformation.IndexRetriever keyInformation,
                             BaseTransactionConfig config,
+                            Configuration disConfig,
                             Duration maxWriteTime) throws BackendException {
         Preconditions.checkNotNull(index);
         Preconditions.checkNotNull(keyInformation);
@@ -65,6 +78,9 @@ public class IndexTransaction implements BaseTransaction, LoggableTransaction {
         Preconditions.checkNotNull(indexTx);
         this.maxWriteTime = maxWriteTime;
         this.mutations = new HashMap<>(DEFAULT_OUTER_MAP_SIZE);
+        // Initialize Dead Letter Queue
+        this.dlq = initializeDLQ(disConfig);
+        this.indexName = GraphDatabaseConfiguration.INDEX_NAME.getDefaultValue();
     }
 
     public void add(String store, String documentId, IndexEntry entry, boolean isNew) {
@@ -139,8 +155,6 @@ public class IndexTransaction implements BaseTransaction, LoggableTransaction {
         indexTx.rollback();
     }
 
-
-
     private void flushInternal() throws BackendException {
         if (mutations!=null && !mutations.isEmpty()) {
             //Consolidate all mutations prior to persistence to ensure that no addition accidentally gets swallowed by a delete
@@ -171,7 +185,7 @@ public class IndexTransaction implements BaseTransaction, LoggableTransaction {
 
                 try {
                     log.info("Calling handleMutationFailure...");
-                    index.handleMutationFailure(mutationsToExecute, e);
+                    handleMutationFailure(mutationsToExecute, e);
                     log.info("handleMutationFailure completed successfully");
                 } catch (Exception handlerEx) {
                     log.warn("Failed to handle mutation failure", handlerEx);
@@ -209,5 +223,106 @@ public class IndexTransaction implements BaseTransaction, LoggableTransaction {
 
     public void invalidate(String store) {
         keyInformation.invalidate(store);
+    }
+
+    /**
+     * Handles mutation failure after all retries have been exhausted.
+     * This is called by the transaction layer when BackendOperation.execute() fails.
+     * Writes the failed mutations to the configured Dead Letter Queue (DLQ) if enabled.
+     *
+     * @param mutations The mutations that failed after all retry attempts
+     * @param cause The exception that caused the final failure
+     */
+    private void handleMutationFailure(Map<String, Map<String, IndexMutation>> mutations, Throwable cause) {
+        log.error("Mutation failed after all retries exhausted for index '{}'", indexName, cause);
+        log.info("handleMutationFailure called - DLQ status: dlq={}, isEnabled={}",
+            dlq != null ? "initialized" : "NULL",
+            dlq != null && dlq.isEnabled() ? "true" : "false/null");
+
+        if (dlq == null) {
+            log.warn("❌ DLQ is NULL! Failed mutations will NOT be written to DLQ.");
+            log.warn("This means DLQ was not properly initialized during ElasticSearchIndex construction.");
+            log.warn("Check startup logs for DLQ initialization messages.");
+            log.warn("Ensure 'index.search.elasticsearch.dlq.enabled=true' is set in your JanusGraph configuration.");
+            return;
+        }
+
+        if (!dlq.isEnabled()) {
+            log.warn("DLQ is initialized but not enabled. Failed mutations will not be written.");
+            return;
+        }
+
+        try {
+            // Try to determine which store failed from the mutations
+            String storeName = mutations.isEmpty() ? "unknown" : mutations.keySet().iterator().next();
+            log.info("Writing failed mutations to DLQ for stores: {}", mutations.keySet());
+            dlq.writeToDLQ(indexName, mutations, storeName, cause);
+            log.info("✅ Successfully written failed mutations to DLQ for stores: {}", mutations.keySet());
+        } catch (Exception dlqEx) {
+            log.error("❌ Failed to write to DLQ after retry exhaustion", dlqEx);
+        }
+    }
+
+    private ElasticSearchDLQ initializeDLQ(Configuration config) {
+        log.info("========================================");
+        log.info("Initializing Elasticsearch DLQ...");
+        log.info("========================================");
+
+        // Log the full config path for debugging
+        log.info("Checking config key: {}", DLQ_ENABLED.toStringWithoutRoot());
+
+        // First try to get from config
+        boolean dlqEnabled = config.get(DLQ_ENABLED);
+
+        log.info("DLQ enabled: {}", dlqEnabled);
+
+        if (!dlqEnabled) {
+            log.info("Elasticsearch DLQ is DISABLED");
+            log.info("To enable, set: index.dlq.enabled=true in your configuration");
+            log.info("========================================");
+            return null;
+        }
+
+        try {
+            log.info("DLQ is ENABLED, loading Kafka configuration...");
+
+            String bootstrapServers = config.get(DLQ_KAFKA_BOOTSTRAP_SERVERS);
+            String topic = config.get(DLQ_KAFKA_TOPIC);
+
+            log.info("DLQ Configuration:");
+            log.info("  Bootstrap Servers: '{}'", bootstrapServers);
+            log.info("  Topic: '{}'", topic);
+
+            if (bootstrapServers == null || bootstrapServers.isEmpty()) {
+                log.error("❌ DLQ is enabled but kafka-bootstrap-servers is not configured. DLQ will be disabled.");
+                log.error("Please set: index.dlq.kafka-bootstrap-servers=<your-kafka-servers>");
+                log.info("========================================");
+                return null;
+            }
+
+            // Additional Kafka configuration can be passed as empty map for now
+            // Can be extended later if needed
+            Map<String, Object> kafkaConfig = new HashMap<>();
+
+            // Use singleton DLQ manager to ensure DLQ remains open for application lifecycle
+            DLQManager dlqManager = DLQManager.getInstance();
+            ElasticSearchDLQ dlqInstance = dlqManager.getDLQ(bootstrapServers, topic, kafkaConfig);
+
+            if (dlqInstance != null) {
+                // Increment reference count for this DLQ instance
+                dlqManager.incrementDLQReference(bootstrapServers, topic);
+                log.info("✅ Successfully initialized Kafka DLQ via DLQManager!");
+                log.info("DLQ reference count: {}", dlqManager.getReferenceCount(bootstrapServers, topic));
+            } else {
+                log.warn("⚠️  DLQManager returned null DLQ (possibly shutting down)");
+            }
+            log.info("========================================");
+            return dlqInstance;
+
+        } catch (Exception e) {
+            log.error("❌ Failed to initialize DLQ. DLQ will be disabled.", e);
+            log.info("========================================");
+            return null;
+        }
     }
 }
