+++ janusgraph-core/src/main/java/org/janusgraph/diskstorage/util/MetricInstrumentedIndexProvider.java
@@ -123,11 +123,6 @@ public class MetricInstrumentedIndexProvider implements IndexProvider {
     }
 
     @Override
-    public void handleMutationFailure(Map<String, Map<String, IndexMutation>> mutations, Throwable cause) {
-        //do nothing
-    }
-
-    @Override
     public boolean supports(final KeyInformation information, final JanusGraphPredicate janusgraphPredicate) {
         return indexProvider.supports(information, janusgraphPredicate);
     }
