+++ janusgraph-solr/src/main/java/org/janusgraph/diskstorage/solr/SolrIndex.java
@@ -1264,11 +1264,6 @@ public class SolrIndex implements IndexProvider {
         }
     }
 
-    @Override
-    public void handleMutationFailure(Map<String, Map<String, IndexMutation>> mutations, Throwable cause) {
-        
-    }
-
     /*
     ################# UTILITY METHODS #######################
      */
